backend/requirement.txt
Django==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-rest-auth==6.0.0
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1